Update README.md
Putting in some beginning policies for management of this project. They are liable to change at any time, and are mainly for me to attempt to codify my personal thoughts on how a project should be managed.

I understand these things won't last if this thing picks up steam (plans and first contact, and all...) and that my feelings on such management practices may change as I age and gain experience.

We'll see how these things change as the project develops.